🛠 :: (#765) interest 도메인으로 이동
