+++ jobis-application/src/main/java/team/retum/jobis/domain/interest/spi/QueryInterestPort.java
@@ -1,7 +1,9 @@
 package team.retum.jobis.domain.interest.spi;
 
+import team.retum.jobis.domain.code.model.CodeType;
 import team.retum.jobis.domain.interest.dto.response.InterestResponse;
 import team.retum.jobis.domain.interest.model.Interest;
+import team.retum.jobis.domain.student.model.Student;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,4 +15,6 @@ public interface QueryInterestPort {
     List<Interest> getAllByStudentId(Long studentId);
 
     List<InterestResponse> getByStudentId(Long studentId);
+
+    List<String> getAllByStudentAndCodeType(Student student, CodeType codeType);
 }
